refacto: extract loadRecordExtractor()
