+++ tests/test-on-html-file.js
@@ -1,26 +1,12 @@
 const fs = require('fs');
-const { URL } = require('url');
 const cheerio = require('cheerio');
 const junit = require('junit');
+const { loadRecordExtractor } = require('./loadRecordExtractor');
 
 const USAGE =
   '$ node test-on-html-file.js <name_of_page.html> [<name_of_config.js>]';
 
 const DEFAULT_CONFIG_FILE = 'config.splitting.js';
-
-const DEFAULT_START_URL = 'http://www.example.com/';
-
-function loadRecordExtractor(configFile) {
-  const { startUrls, actions } = require(`./../${configFile}`);
-  const { recordExtractor } = actions[0];
-  const defaultUrl = new URL((startUrls || [DEFAULT_START_URL])[0]);
-  return ({ url = defaultUrl, title = 'Some page', html = '' }) =>
-    recordExtractor({
-      url,
-      title,
-      $: cheerio.load(html),
-    });
-}
 
 function test(recordExtractor, htmlFile) {
   const it = junit();
